requirements.txt
spacy
ipykernel
pandas
numpy
matplotlib
seaborn
black
flake8
scipy
statsmodels
nltk
google-cloud-speech
ipython
pytextrank
pydub
pre-commit
dask